Add files via upload
It's a mini project. It can be helpful for course projects. Any University or college have many clubs. If there is a available website for clubs, then anyone can visit the website anytime. Here is an example. Contents - (Home, About, gallery, Election committee, Clubs-some clubs pages, notice).
Here is a drive link for this project also -
https://drive.google.com/drive/folders/1SiEaOmw7gSRdmQ70E8OTK1653nLeFeJV?usp=share_link